Follow up of #28

This is the final place where a RWMutex was being using the same approach (immutable cache + atomic switch for updates).

Here are the numbers compare master and this branch:
```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/consul-go
BenchmarkBalancer/LoadBlancer+RoundRobin            	30000000	        56.5 ns/op
BenchmarkBalancer/LoadBlancer+RoundRobin-8          	30000000	        56.6 ns/op
BenchmarkBalancer/LoadBlancer+RoundRobin-32         	30000000	        56.4 ns/op
PASS
ok  	github.com/segmentio/consul-go	5.270s
```
```
goos: darwin
goarch: amd64
pkg: github.com/segmentio/consul-go
BenchmarkBalancer/LoadBlancer+RoundRobin            	50000000	        35.4 ns/op
BenchmarkBalancer/LoadBlancer+RoundRobin-8          	50000000	        35.1 ns/op
BenchmarkBalancer/LoadBlancer+RoundRobin-32         	50000000	        36.4 ns/op
PASS
ok  	github.com/segmentio/consul-go	5.475s
```
```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkBalancer/LoadBlancer+RoundRobin        56.5          35.4          -37.35%
BenchmarkBalancer/LoadBlancer+RoundRobin-8      56.6          35.1          -37.99%
BenchmarkBalancer/LoadBlancer+RoundRobin-32     56.4          36.4          -35.46%
```

In this case the use of RWMutex wasn't much of a problem since the inner cache of the load balancer is rarely updated so it almost never blocks, but the resulting code is still much faster.